Add basic UI
Add a way to visualize and inspect the current image pipeline.

Previously we just assumed all objects were composited together. This
will not be true going forward. E.g. if we execute a cut and paste,
there will be intermediate masking objects that we may want to inspect,
but not include in the final output

Add a composition object, who is responsible for layout of other
objects. Add a basic UI that lets us look at any object (including
composition objects)

UI is currently using imgui, this is likely to change in the future but
is easy to get up and running