Aspect ratio/zoom/pan
Add viewport coordinate transforms that handle zoom/pan and aspect
ratio.

The new behavior is that the expected "default" transform ensures that
objects are scaled in x/y such that the aspect ratio is maintained. The
default behavior is to fill the window/composition in whichever
dimension fits best.

Note that within objects, the coordinate space has not changed. The only
change is about the expectation of how objects will be positioned in
eachother. This definition extends to compositions, where the transform
now takes the aspect ratio adjusted object and positions it within the
composition, which has been given a fixed size for layout

With this transform, we can also implement pan/zoom. This was initially
done in multiple stages, but ultimately the introduction of these
abstractions is better justified when pan/zoom is brought in at the same
time

Panning is triggered with middle click drag, zoom is triggered with
the scroll wheel

Implementation notes:
* Transform type has been refactored to be separated from a general 3x3
  matrix. This allows for some nice chaining of transforms
  a.then(b).then(c) instead of having to matmul right to left
* The view state is responsible for handling the transformation from
  window space/clip space to object space.
* Path rendering now needs to take in a transform
* Pixel dimensions were being taken in/out from many places, add as a
  type in lin.zig
* Matmul had to be implemented to stack transforms, fix some hacks
  now that we have it