Split input state from object storage
Previously we were storing what is essentially UI state within an
object. E.g. which part of a path was selected on mouse down was stored
in the path. This info is not really part of the object, but more part
of the UI state. If we have multiple paths, only 1 can have a selected
object, but we were storing the selection in each path

As we add more UI features with more state, this problem will only get
worse. Fix it now to avoid a bigger pain down the road. Split out an
InputState struct which holds state related to future input events.

This pushes the App struct into essentially a middle man between object
storage, view state, input state, and rendering, which is kinda nice

Automatically convert glfw keys to character codes and use those in the
input state mappings