Add support for scaling items in composition view
When the s key is pressed in the composition view, lock the cursor to
wherever it is on the selected object. Scale the object to keep the
location on the mouse. Scaling can be aborted with a right click, and
applied with a left click

Input state caches the transformation of the object when the scale was
started, and updates the transform with the new scale every frame. Scale
is provided as the first item as we want scaling to always stretch the
object in its own local x/y

Transform inversion had to be implemented to convert the cursor
position from composition object space to composed object space.