Fix panning
View center was supposed to be in object space. Previously the view
center was in clip space, but in
c6ec3fec437250506210211a27f47261d5dd23bf we changed it to be in object
space.

This means that we have to change the order of operations when
generating the viewport transform.

While we're here, use the inverse of the transform in clipToObject, and
rename objectToWindow to a more correct windowToClip