Add UI to create compositions
Create a "create object" panel, which holds all the ways to create new
objects. At this point in time this is only paths/compositions, but in
the future this will contain more

The path creation button has moved here, which makes it less clear that
the path is tied to the current selected object. Keep the selected
object as the default display object, but allow editing after the fact

Editing generated paths makes it very easy to create dependency loops.
Add a dependency loop checker and ensure that modifying objects does not
create loops.

Loops are still shown in UI. This is because...
1. Checking if a loop would be created by changing something is a lot
   harder than just doing it and checking. AFAICT the easiest thing to
   do would be to duplicate object storage and do the check on the
   duplicated tree after making the modification. Just trying and
   rolling back is way easier.
2. Hiding objects depending on what creates a loop could be confusing.
   Letting the user try something before rejecting it makes it clear
   what they need to do to make it work

Since failure is a semi-expected case now due to loop detection, fail
gracefully in these cases instead of crashing the app

Now that dependency loop detection is available, there's no reason to
restrict adding a composition to a composition, remove that restriction

Bonus fix to lint program to try deleting objects, noticed while
coverage checking lint program for new UI apis