Add UI to change shader input objects
Shaders act on other objects. For the time being we're hardcoding a
color swap shader for each input image as a proof of concept, but in the
future we expect users to be able to instantiate whatever shaders they
want, for whatever images they want.

Add UI to adjust the inputs to give users more control