Defer shader dims to one of the input uniforms
Shader objects need to report some resolution. At the very least we need
an aspect ratio for reasonable layout in the window/composition.

Manually setting this in the UI would suck. If we change the input
texture from one object to another, the input dimensions will change as
well.

Instead forward the resolution to match one of the shader inputs. Allow
the user to select which one