Fix shader -> shader runs
Our framebuffer context used to unconditionally disable the framebuffer
on reset. In some cases we want to render some shader object to a
texture, and in the process we discover that we need to render _another_
shader object to a texture. In this case we double nest framebuffers.

Restore whatever the previously bound framebuffer was