Add "reserved" uniforms to program runner
In a future commit, we will be injecting distance fields into our
brush shaders. We do not want these to show up in the UI, so we need
some way to inform the system to consider some uniforms user editable,
and some not

Add the concept of a reserved uniform. These are uniforms that will be
modified by the system. As of right now there is only the aspect
reserved uniform used by the checkerboard shader.

Added user modifiable checkerboard shader as example