Text rendering impl
Putting text in an image is a key component of any good image editor.
With a large dose of NIH, we have implemented a font renderer ourselves.

This change adds a poor man's TTF renderer, and a text object to use it.
Anti-aliasing is pretty poor, but it can get the job done. We use a
distance field + a shader to feather the edges of the font, which seems
to be sufficient for it to be legible at low resolutions

At this point we do not track the letter positions at all. We just stamp
the glyphs next to each other, starting at the bottom left corner. This
is obviously not good, but it is funny, so it's a good checkpoint for
the initial implementation

Lots of empirical testing done here. I've been sitting on this change
(or a variant of it) for a week or so. At this point I'm willing to
commit it with a less than thorough review since there's been some
natural hardening time. I did look, but there's a lot of code...

Implementation details:
* Pull in ttf prototype and adjust for our usage. There's a lot of code
  here, but not much to talk about.
  * Mostly referencing the ttf specs from apple and microsoft, doing
    whatever we have to do to get fonts from Hack rendered
  * Rendering is done as a signed distance field, this is generated in
    mega big space. 1 pixel per funit, then is scaled down to be put in
    a glyph atlas. The high res is to ensure high accuracy of the sign
    part of the signed distance field
* Added font storage to App, text objects reference fonts with a font ID
* GlyphAtlas abstraction added to store pre-rendered glyphs
* TextRenderer abstraction feeds a glyph atlas and generates a series of
  quds to render from the atlas
* Updated distance field renderer to support generation of signed
  distance fields if a mask is passed in
* Added support for non rgba32f texture generation... masks don't need
  it, and can be bigger without
* Committed a couple ttf files for use in linter script
* TextObject added. This calls the TextRenderer and uses it to layout
  text from a font into an image

Bonus change:
* Unit tests in linter script
* Incomplete framebuffers do not crash the app, they just complain
  loudly