Add short term object to hold render state
Passing around all the app state to every render function is annoying,
and makes diffs with new state hard to parse. Split out an object that
holds all state for a single frame. This object will also be used in the
future to pass around a single parameter to other objects that want to
render stuff