Fix dragging in composition view
In some cases (nested compositions at least) our movement when dragging
items in a composition was totally wrong. This is caused by some
unconventional coordinate spaces.

Previously we tried to work around this by applying the transformation
to the start and end pos. The thought must have been that this would go
through all the necessary parts of the transform, and then we could
apply that first.

This is wrong. It seems pretty clear to me looking at this now that the
transformation should happen as the last step. This is infeasible
however due to the aspect ratio correction.

Ok, do it one step back from the last step and just compensate for the
future apsect correction