Split out some modules
This is motivated by a future commit where we want to use the text
renderer as part of our incoming gui framework. With the amount of stuff
in this upcoming change, it would be nice to be able to separate out a
gui folder and throw all our widgets in there

If we want to render text in the gui framework, we need the text
renderer. If we're using the text renderer, we need Renderer.zig, which
pulls in Object.zig, etc. Everything is so tied together we essentially
have to put everything in the root level. This makes
navigation/organization difficult

Split out modules as necessary to enable using the PlaneRenderProgram
and TextRenderer. Try to categorize stuff at least a little bit

Changes...
* UniformDefault is now a distinct type from UniformValue. This is
  because our UniformValue injects an ObjectId, which is reasonable. The
  PlaneRenderProgram only uses ResolvedUniformValues, and we only need
  UniformValue in any dependencies to resolve defaults. Since we can't
  have a default ObjectID, this is a good candidate
  * While we're moving stuff, we might as well split out the
    PlaneRenderProgram and DistanceFieldGenerator into their own files
* lin.zig + parts of coords.zig moved into new sphmath module
* geometry.zig, gl.zig, and parts of Renderer moved into new sphrender
  module
* GlyphAtlas.zig, TextRenderer.zig, and ttf.zig moved into new sphtext
  module
* build.zig touched pretty heavily to make appending all these modules
  easier