Text rendering improvements
* Implement text wrapping
* Fix trailing spaces

Text wrapping for now is simple. Write down the start of each word. If
the laid out text hits the wrap boundary, advance to the next line and
re-layout starting at the start of the word again. If we hit the edge
again, split the word

Layout complexity has increased to the point where it is worth splitting
out a LayoutHelper class which moves character by character