Fix text rendering to support running with glScissor enabled
In some future gui code, we use glScissor to enforce that widgets do not
draw in places they are not supposed to. This breaks our text rendering
code as we assume that no scissoring is enabled. Ensure the glScissor()
contains the bounds of the entire render area when rendering our
distance fields and glyphs