Replace imgui with our own gui framework
Implement a gui framework from scratch. This is done mostly so that we
can eventually say we have no dependencies other than OpenGL. Now is a
good time to it. Our font rendering is pretty good, and we're starting
to look at implementing some custom widgets/UI behavior (e.g. adding a
thumbnail drawer for object selection)

While imgui was a fine prototype UI, it was never intended to be used
long term. I do not want to pour work into figuring out how to make
custom widgets with it if we know we are going to swap it out later. If
we make the UI ourselves, we also can design the UI in any way we
imagine, without needing to take restrictions of some third party
dependency into account

The GUI framework is designed around the pattern that I've ended up
using in many projects in the past. The UI subsection of the code should
be a read only view into some app state. When actions are performed in
the UI, the UI should request those actions to be performed, and the
main application can perform those actions as it needs

We have taken some inspiration from immediate mode GUI frameworks.
Specifically the idea that we can re-layout every frame, and that is
often simpler than trying to build some form of complex notification
system between components.

The immediate mode idea falls over a little when you start considering
inter-frame state though. If a widget wants to track position over time
it needs somewhere to put the state. E.g. input debouncing, animations,
holding last frame's text layout to avoid expensive relayout. So instead
of going completely immediate mode, we have a widget framework that
works around the idea of "retrievers". Retrievers are passed into
widgets on instantiation as a way to instruct them how to get the latest
state from the application. E.g. a label could take in a retriever with
a `fn getText() []const u8`, which will return different strings
depending on the current app state.

The core component of the GUI framework is the Widget. Widgets have a
fairly low API surface area, but the core of it is

* update(): Fetch data, regenerate internal widget state, expand to fill
  new widget bounds, etc.
* setInputState(): Handle input, and return an optional action to be
  performed by the app
* render(): Take data from update(), and put it on the screen

The initial state of the UI is meant to mirror the existing UI as best
it can. We keep the 3 pane layout with a create view, and object list,
and a property list. Some details about those views may have changed,
but the intent was to start with a 1-1 replacement, then tack on fancier
new features later.

This is a major commit, containing weeks of work. Typically I would try
to commit in a more incremental fashion, but in this case tracking the
intermediate work is not valuable. The GUI framework wasn't so much
designed as it evolved. The intermediate states were ugly,
non-functional, and not very coherent. For this work it makes more sense
to commit the final thing, rather than all the mistakes we made along
the way

Due to the way the UI is hooked up now, some of the APIs to the
application changed. E.g. some containers now need to be queried for
numItems(), or the GUI framework now asks for single components of vec3s to
be updated in shaders. Most of these changes are fairly simple to grok,
but some extra notes are provided below

App
* Some functions are now public to assist with querying info related to
  the currently selected object
* The GUI framework now exports lower instead of upper characters on
  keypress, so some input state ascii checks needed to be updated

Objects
* Shader editing is actually very similar between drawing and shader
  object types. New API added to extract current uniform values from
  whichever one we are using

Rendering
* glViewport management is no longer done by the App renderer. This
  conflicted with some of the expectations of the gui framework. This
  means that the FrameRenderer no longer needs to track the window size
  in pixels

TextRenderer
* empty layout added for gui_text widget initial state
* layout adjusted to have a consistent height no matter what
  characters are inserted
* pixToClip types changed since all our inputs are u32/i32 usize is a
  strange choice there