Convert submodules into importable packages
It turns out that some of the modules we have written are useful
elsewhere. We could split these modules out into their own
repos/libraries etc, but in reality they are closely tied to this
project. Allow others to use them if they want, but keep them as first
class citizens of sphimp