Fix missing - in rendered text at low point sizes
Off by 1 error in the glyph atlas allocation. We have a 1 pixel border
between glyphs to allow for slight inconsistencies when sampling the
atlas. This padding was accidentally double applied on the y axis. We
probably thought "we subtract here in x, why not y?". On the x axis we
calculate the right edge by the new cursor position, where on the y
axis the cursor doesn't move, so we don't have to offset the padding