Improved composition object selection
Previously object selection felt pretty janky. We always selected the
object who's center was closest to the mouse click position. This made
prediction of which object would be selected very difficult.

Instead we should check if there are relevant pixels of the object shown
where we click. This is done by rendering composition idxs into an integer
texture, then sampling that texture to see which one we hit. We use a
texture size of 1x1 so there should not be that much rendering to do

Debug visualizations for the hit testing were quite useful while
implementing. We didn't have a checkbox widget before, so one was added
to toggle the debug view

Implementation details:
* Add checkbox widget
* Render object mask when generating composition transformation state in
  input handler
* Update uniform handlers to support uint params
* Disabled mipmap generation completely. AFAIK, we don't want mipmapping
  anywhere anyways, and were just doing it for safety. Integer textures
  can't be mimapped, so we could either special case them and leave
  floating point textures in an undesirable state, or we could fix it
  all