Move shader transform into generic API
When implementing integer backed textures (in a future commit) we did
not notice that the system did not know how to apply uniform sampler
textures. This was caused by a combination of two design decisions

* Some shaders have uniforms that we do not know how to display to
  users. E.g. Mat3x3. In this case erroring would be wrong so we just
  choose to ignore parameters we do not understand
* Glsl will compile out uniforms that we expect to exist if they happen
  to be unused. If we try to set these uniforms and they don't exist,
  that's fine, just ignore it

When you combine these two decisions, you end up with a silent failure
to work with types you care about

When applying image uniforms, the active texture is assigned, and we
bind a texture to that active unit. In the case where we use a sampler
type we do not support, we will just use whatever texture happens to be
bound.

Why? The default texture unit is 0, and we will likely have texture unit
0 bound due to a previous attempt to set a image texture (or maybe its
bound by default, idc enough to look it up). This means that we
confusingly just show a random texture instead of the one we want

So we need to change one of the two design decisions above so that we
can notice when we are doing something stupid.

If we set transform (mat3x3) uniforms through the standard API, we no
longer have any built in shaders that have unknown params. This allows
us to log a warning when we encounter a new thing we don't support

Implementation details:
* Remove transform from the PlaneRenderProgram API
* Set transform through standard ReservedUniform API
* In a few places we do conversions between different related uniform
  types (Default, Resolved, sphimp uniform). Most of these conversions
  are passthrough backed by the same UniformType. We can use some zig
  compile time magic to auto-propagate
* Add transpose() to sphmath Mat3x3 because reading default values from
  opengl are column major, our storage is row major
* sphmath unit tests were previously not being run (oops). And even if
  they were, the tests nested in structs were not being run (oops)
* GlyphAtlas is always rendering with an identity transform. We can make
  the transform optional if we set identity by default in the shader