Improve shader API
Our original PlaneShaderProgram evolved from a specific program that
sampled a rectangular image onto a rectangular plane with a specific
transform. Eventually this was modified to support arbitrary user
provided shaders and eventually non-square shapes.

At this point we really want a generic OpenGL abstraction. Replace
the PlaneRenderProgram with a program that can render an arbitrary
buffer type with arbitrary uniforms.

The major reasoning for these changes is...
* Reduce friction for working with vertex shaders that do not follow the
  typical XY + UV + transform
* Improve API for working with uniforms known by the program

Uniforms are now defined by a structure, not an index uniform. This
means at compile time we can automatically derive which OpenGL calls to
make for a specific field, and usage of setting uniforms feels very
natural and zig like. E.g.

```
program.render(buf, .{ .transform = transform, .color = color });
```

Can set the transform and color uniforms, as long as we've defined a
struct with the appropriate transform/color types

Vertex attribute offsets are now also automatically determined by an
arbitrary structure. The new shader program tries to match names of
fields in structs with inputs in the shader program, and sets the
attributes according to the offsets within the struct

The new API is expected to be used mostly in contexts where all program
uniforms are known by the program. This means that the case of iterating
user modifiable uniforms is the rare case, and should be opt in instead
of opt out. Move uniform discovery into a fn call thats only called for
shaders stored in ShaderStorage

Replace all usages of PlaneRenderProgram with the new type. Other types
could be replaced as well, but I don't feel like handling those changes
at this point in time

While we are here, fix hacks around re-using the same vertex buffer
between multiple programs. I'm unsure that this is valid, and the new
API coerces us into doing the more correct thing and associating
specific vertex buffers with specific programs