Stack widget rework
We used to have 3 layout modes, center, fill, and offset. These
controlled how widgets were positioned in the stack, and what sizes they
were provided during update()

This is not sufficient for the upcoming implementation of the image
drawer. Our approach here is to position the tab above the widget to the
left of the drawer. As the stack size changes, we need the tab to stay
centered vertically, but clamped to the right edge so that it looks like
it's attached to the drawer.

Rework the layout to have separated horizontal justification, vertical
justification, and sizing.