Image drawer implementation
It has become apparent that working with images by name is not very
easy. Add a visual image drawer to navigate by thumbnail instead

Several new supporting types have been added
* clickable - widget that converts any widget into a button
* grid - layout items in a grid, updating downstream widgets with a
  proportion of the available size
* thumbnail - put a picture on the screen
* image renderer - Formalize the use of xyvut as an image renderer so
  that we can share it between widgets if necessary
* xyt - like the xyuvt widget, but without uv. For constant color
  shaders

Overall setup

The image list is stored in a "drawer", this drawer is usually hidden,
but can be opened with the ~ key, or by clicking on the tab. The tab is
positioned on top of the right edge of the app widget.

When the drawer opens it slides in from the right. You can click
thumbnails in this view to select an image

Drawer is constructed out of a scroll view, that holds a grid, that
holds, a clickable thumbnail clamped vertically to prevent it from
eating the whole screen

Supporting changes
* The grid does not know the height of a row without checking the height
  of each element in the row. This means we need to iterate from N..N+M.
  Add iterFrom to our segmented list