Make app installble
If this is to replace our typical usage of a paint app, we need to be
able to easily launch and start drawing. Add the appropriate items to
the zig build output to launch

Install all shaders/ttfs/brushes and load them off disk on init. This
allows us to launch without a convoluted command line