Fix error prone allocator passing in object data
Each object owns its own allocator, since this is true of all objects,
it is owned by the Object structure. Specific object data also needs
this allocator, but previously we had not found a simple way to find it
without passing the allocator to every function

With @fieldParentPtr we can remove the need for allocators to be passed
in. Do that