Add eraser to drawing view
Add the concept of tools to the drawing view. As of right now there are
only two tools, the brush and the eraser.

Erasing is done by splitting a stroke at a given point. We walk all
points within some range of the cursor and remove them one by one. This
is not optimal, but never seems to be noticably slow. If it becomes a
problem we can always try to batch a little more intelligently

Supporting changes:
* Add support for rect to use a color retriever. This allows us to
  implement the highlight for our clickable widgets with a rect + stack
* Add one_of widget. This will toggle between 1 of many sub-widgets, but
  uses the maximum size of them. This allows us to toggle between
  different params lists for different tools without the scroll view
  jumping around
* Make drag action optional in interactable, as we don't need drag on
  the tool view
* Time added to app.render() so that we can preview the eraser size for
  some period of time
* Tool params have been added to App. Parameters for each tool live for
  the lifetime of the app, even if they are not currently in use. This
  makes flipping back/forth between tools feel more natural as they
  don't reset
* Add ui renderer for overlay/preview stuff
* Add StbImage variant that works on raw data instead of a path. This to
  load embedded image data