Fix focus not propagating correctly in grid view
The grid view removes the focused widget id when focus is set to false,
the layout was setting focus to false every frame, then setting it back
to true. Due to the set to false, we lost the focused id and could not
put it back. I believe the grid view behavior is sane, so fix the
calling code

Introduced in b185f75e91946f085c6f2aa8f8b97e3a36e372af