Edit dependencies of an object without changing viewport
Often we want to see how changes in one area effect some composition.
With the existing UI, we had to toggle back and forth between elements
quickly, zoom in, make our change, switch back, etc. etc.

Move towards allowing editing any attribute of any dependency without a
change. For now this is restricted to items that can be changed in the
property view, but in the future could (and probably should) be extended
to any action at all

Navigating the hierarchy of objects is not a simple task. The dependency
graph can explode in complexity very easily. We considered a node based
system, but did not like how messy those can get. We experimented with a
few ways to show all dependencies at once, but none of them were
satisfactory. The closest attempt was a radial view where each level of
the dependency tree gets further from the center, and smaller as well.

We settled on a new style of widget. Right click in the top left window,
and drag the mouse left and right to flip through children of the
current selection. Drag the mouse down to add it to the selection stack,
or up to go up a layer. This feels pretty nice so far, future
improvements may come

Implementation details
* Frame size underflow fixed with a saturating sub. I don't think this
  is required anymore, but I'm not sure what triggered it so we leave it
  in
* Main loop can now independently change the "property object" and the
  "selected object". One is for the sidebar, one is for the main
  viewport
* A pattern that keeps coming up is that we have some unknown amount of
  stuff. Pre-allocating our super big unhittable number is silly, as we
  likely will never need that much space. Instead we use the
  RuntimeSegmentedList. But what if we want to return an array at the
  end? Add a makeContiguous API that copies each segment of the list
  into some output allocator