Remove dependency on mouse position in frame renderer
This did not belong. Frame renderer is used in many contexts, most of
which should not render UI elements. E.g. thumbnails, image export, etc.

When we remove input_state from App, callers of frame renderer
will not even have access to the mouse position. This is now
inconvenient on top of being wrong, so we fix it.