Separate long term and short term frame renderer allocations
Some users of the frame renderer use it to create long lived items (e.g.
thumbnail caches). We expect to leak those long term allocations, but it
is not expected that short term allocations are leaked along with them.
We should only be using the memory that we are actually using

Feed scratch allocator into frame renderer so that short term allocs can
be short. Feed in as std.mem.Allocator, not sphalloc.Scratch because in
many cases we use the scratch allocator for both long and short term
allocations