Add support for hiding cursor
In a future widget, we are planning on hiding the cursor and locking its
position. Imagine it is being used as a joystick in this context.

The widget needs to be able to request that the cursor is hidden. We
cannot do this as a typical ui action, because we may want to disable
the mouse _and_ do something else. Add cursor style as a gui
InputResponse output

Delete unused even_vert_layout, since it isn't used and its a little bit
of work to update the setInputState API