Add function to copy backwards on scratch allocator
A common pattern is to do some processing of data in a scratch
allocator, and return the results in a second allocator before resetting
the scratch space.

In some cases though, we may want to return the output in the same
scratch allocator. In this scenario we need to allocate memory from the
already allocated scratch space, copy backwards without destroying the
previous content, and set the new scratch buffer length

Add a function to do so