Add async library + HTTP server
What is async, if not a state machine stapled on top of epoll? Add an
event loop that is capable of running an http server

APIs may feel a little odd. They are just whatever felt right when
writing a real webapp. Things may be missing, etc. etc. This is just a
good starting point