Add way to extract contiguous slice from RuntimeSegmentedList
Lots of APIs expect contiguous slices. Sometimes we might want to pull
out a subset of a large segmented list and use it in these APIs.
Typically it should be free, but sometimes we need to allocate a little

Bonus change: Add an allocator that always fails which API consumers can
use if they want to be sure they aren't allocating