Clear scratch memory on free
Leaving everything we allocate around and available to mistakenly use is
just asking for memory issues

This also exposed a bug in restoreCheckpointAndPreserve where we
actually were not restoring to the checkpoint at all :). This was
supposed to be an optimization for copying some data at the end of the
allocator towards the beginning of the buffer, but instead we were
actually wasting more space than if we had done nothing at all :).
Oops.