Add shutdown support to event loop
"Common" (single consumer lol) use case is to run a server until SIGINT.
This requires us to wake up the event loop and close all connections.

Add a signalfd abstraction and a way to call close on all handlers in
the event loop

During the implementation of the signalfd abstraction we discovered
passing in comptime ctxs feels pretty nice. Give the same treatment to
the Server type to assist with testing