Add HttpConnection to event library
Generic http connection. Calls serve() on some context who is in charge
of writing back over the tcp connection with it's own HttpWriter

sphevent library now depends on sphttp. This made more sense than doing
it the other way around, and then spinning up a new lib. sphttp is a
cheap dependency :).

Ideally we would write a unit test for this but I was feeling lazy :)