Scratch allocator improvements
Sometimes a function wants to return data on the scratch allocator while
also using some intermediate scratch space internally. This can lead to
using more memory than necessary as the intermediate scratch space
cannot be reused until the scratch space is reclaimed.

This can cause real issues when there are long loops being executed
between reclamation

Update the scratch allocator interface to support allocating from both
ends of the scratch space. This allows functions to clear different
types of work independently. E.g. We can put return data on the front
side, scratch work on the back, and restore a back checkpoint before
returning.

The default assumption is that we allocate from the front of the buffer
(or the side with smaller indexes). This allows us to act as a mostly
drop in replacement for the previous ScratchAlloc. Some ScratchAlloc
functions have not been reimplemented, and some likely will never have
to be as their use case is replaced with the back allocator