Fix crash on shrinking empty RuntimeSegmentedList to empty
Our unused block iterator determines the first unused block by finding
the last block that _is_ used and adding 1. I.e. getBlockFor(list.len -
1). If list.len == 0 this underflows.

Replace - with saturating sub, this results in len 0 returning block 0,
which gets incremented to 1. This is correct because block 0 is the
pre-allocated block which we should never free