Histogram visualizer
Add way to visualize histograms. This will probably come up more than
once, so making it as a builtin widget feels reasonable

It's likely that this could with a little bit of work turn into a
general bar chart. It already is not _really_ a histogram (y values are
f32s lol, x values are literally any text). But we call it a histogram
because that is how we are currently using it

Meant to be fairly abstract. Data format is up to user and we provide a
retriever with a few APIs to pull out the data we expect. X labels can
be minorly expensive to generate, but y value getters are expected to be
cheap.

To avoid re-generation spam retriever must provide a "data generation"
which increments every time we want a redraw

Details...
* Add histogram with bell curve to demo app
  * Previously raw label is now layout of label + histogram
* Fix for zig 0.14 missed compile errors in gui_text (enum casing)
* Added support for signed integers in GL vertex + uniform abstractions
  * Used for hover index in histogram highlight
* Added support for Vec4s in GL uniform abstraction
  * Used for bar colors