Add multi line graph widget
Add a graph widget that renders many line graphs overlaid on top of
each other.

This was added for the purpose of comparing loss curves for our neural
network training framework. Some decisions that may seem odd here should
be explained by that (e.g. the choice to always put 0, 0 at the bottom
left of the graph)

If other scenarios need to be supported, they can be added later