Properly support scratch RuntimeSegmentedLists
Originally this was designed with the intent of being an expandable list
in the sphalloc system. Initial allocations are done on arena, growth is
done on the sphalloc tiny page allocator to support freeing but without
the cost of using a general purpose allocator

A couple downsides here
* The type was not usable without opting in to importing sphalloc
  because sizes of expansion slots were determined by the tiny page
  allocator size
* When using this type with an arena as our expansion source we would
  waste memory by "freeing" expansion slots on shrink
  * There is no reason for us to not support the scratch allocator use
    case which has turned out to be super useful

Move runtime segmented list into sphutil_noalloc and configure min
allocation size and free support at compile time. This is sufficient to
support both of the above cases

In addition to the above two goals, our implementation also supports
using the real system page allocator as an expansion source if someone
happens to want that

Expose RuntimeSegmentedListLinearAlloc as an arena friendly alternative

Rename existing RuntimeSegmentedList to RuntimeSegmentedListSphalloc to
force downstream users to make a conscious choice on which one they want
on upgrade

Bonus fix:
* Grid gui class was using the block allocator incorrectly