Remove user provided free list constraint from tpa
This was annoying for a couple reasons

1. The type contained the free list size, which meant passing it around
   needed generics or extra type aliases
2. How the heck am I supposed to pick a reasonable size when I make my
   app. I just wrote 100 everywhere and who knows if that was right

Replace tiny page allocator storage with a runtime segmented list backed
by the system page allocator. This allows for an absurd level of growth
that will never be hit at a relatively low cost.

Make initial allocations from a stack buffer contained in the TPA, which
means that we now need an initPinned function

Buddy allocator impl can no longer assume that the free lists are
contiguous, luckily the only place this was used was when scanning free
lists to look for our sibling allocations. Throw an iterator at it

Noticed that we didn't build the demo app in the lint script, which
meant some API changes weren't compiled. Remedy this while we're here