Add a simple hash map
Add a hash map that fits better with our typical allocator strategies.
Explicitly calling out the expected use cases...

1. Long term growable allocations with a pair of arena/block
   allocators. Some long term object has a map that it knows the typical
   size of, but wants to allow growth up to another size. If map entries
   are removed we should release batches of memory back to the system
   without having to do individual frees
2. Linear allocator scratch space. We have a short lived map with a
   known likely size, but an unknown upper bound. Think json parsing for
   a short lived object

In both of these scenarios the standard library hashmap is wasteful of
memory, since we use allocators that do not support individual frees. As
the map is re-allocated we leave behind chunks of memory that we could
have continued to use

Implement a simple hash map based around RuntimeSegmentedLists for both
bucket and node storage. Chunks of buckets/nodes can be dynamically
added/removed up to a certain size

We do not do anything fancy here. Each bucket just is just the head of a
singly linked list

Performance is a little hard to characterize here. Benchmarking 100M
insertions interspersed with 1M removals on a i32 -> i32 map, we see
anywhere from a 50% slowdown to a 10% speedup compared to the standard
library implementation. The 50% slowdown is probably more realistic, but
due to the API forcing us to pick an initial size it is likely that the
map will not be under load. Depending on the page size, we can also end
up having our first expansion being surprisingly large, making our
performance better at the cost of memory usage

Implementation details
* Add fillBlock() function to runtime segmented list. Bucket storage is
  upgraded by just packing as many buckets as fit into each block. We're
  allocating the memory, we might as well use it
* Add indexFromPtr() function to runtime segmented list. This allows us
  to go from node -> storage idx without having to track that index
  explicitly. This gives us a noticeable ~5% speedup
* Add blockStartforIdx() function to runtime segmented list. This allows
  us to rehash when removing nodes. We can see what the size of the
  list would be if we popped the last block, and then shrink to that
  size if it seems like a good idea
* HashMap is built up of a RuntimeSegmentedList of buckets and nodes.
  Nodes are sized according to inputs, bucket storage is 25% larger to
  avoid major hash collision
* Not much else to say, we implement a few APIs that were initially
  useful, we will likely expand on this later